+++ packer-images/scripts/Install-Software-V3.ps1
@@ -1,0 +1,10 @@
+choco feature enable -n allowGlobalConfirmation
+choco install microsoft-edge -y
+choco install git.install -y
+choco install netfx-4.8-devpack -y
+choco install dotnet-6.0-sdk -y
+choco install powershell-core -y
+choco install winrar -y
+choco install notepadplusplus -y
+choco install azcopy10 -y
+choco install heavyload -y